fix: implement signing methods for both impersionated and normal accounts
